moving code into pkg and orchestrating it via main.go
