+++ samples/go/pkg/hmac/hmac_test.go
@@ -1,4 +1,4 @@
-package main
+package hmac
 
 import (
 	"github.com/stretchr/testify/assert"
@@ -7,36 +7,36 @@ import (
 )
 
 func TestGenerateReturnsAnHMACString(t *testing.T) {
-	headers, _ := generate("api_key", "api_secret", "")
+	headers, _ := GenerateHeaders("api_key", "api_secret", "")
 	expectedSignature := "Signature keyId=\"api_key\",algorithm=\"hmac-sha1\",headers=\"date x-mod-nonce\",signature=\""
 	assert.Equal(t, expectedSignature, headers["Authorization"][0:86], "generate should return the hmac headers")
 }
 
 func TestGenerateReturnsADateHeader(t *testing.T) {
 	injectMockDate()
-	headers, _ := generate("api_key", "api_secret", "")
+	headers, _ := GenerateHeaders("api_key", "api_secret", "")
 	expectedDate := "2020-01-02"
 	assert.Equal(t, expectedDate, headers["Date"])
 }
 
 func TestGenerateReturnsANonceHeaderWithExpectedValue(t *testing.T) {
 	nonce := "thisIsTheNonce"
-	headers, _ := generate("api_key", "api_secret", nonce)
+	headers, _ := GenerateHeaders("api_key", "api_secret", nonce)
 	assert.Equal(t, nonce, headers["x-mod-nonce"])
 }
 
 func TestGenerateReturnsAGeneratedNonceHeaderIfNonceIsEmpty(t *testing.T) {
-	headers, _ := generate("api_key", "api_secret", "")
+	headers, _ := GenerateHeaders("api_key", "api_secret", "")
 	assert.True(t, headers["x-mod-nonce"] != "", "x-mod-nonce header should have been populated")
 }
 
 func TestGenerateThrowsErrorIfApiKeyIsNull(t *testing.T) {
-	 _, err := generate("", "api_secret", "")
+	 _, err := GenerateHeaders("", "api_secret", "")
 	 assert.Equal(t, "api_key cannot be empty", err.Message)
 }
 
 func TestGenerateThrowsErrorIfApiSecretIsNull(t *testing.T) {
-	_, err := generate("api_key", "", "")
+	_, err := GenerateHeaders("api_key", "", "")
 	assert.Equal(t, "api_secret cannot be empty", err.Message)
 }
