+++ samples/go/pkg/hmac/signature/signature_test.go
